+++ yyGang/src/main/java/com/beyond3/yyGang/CartOption.java
@@ -1,0 +1,38 @@
+package com.beyond3.yyGang;
+
+import com.beyond3.yyGang.cart.Cart;
+import com.beyond3.yyGang.nsupplement.NSupplements;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import lombok.Getter;
+
+@Entity
+@Getter
+@Table(name = "cart_option")
+public class CartOption {
+    @Id
+    @GeneratedValue
+    @Column(name = "cart_option_id")
+    private Long cartOptionID;
+
+    private int quantity; // 수량
+
+    private int price;  // 가격
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "cart_id")
+    private Cart cart;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "products_id")
+    private NSupplements nSupplements;
+
+
+
+}
