+++ yyGang/src/main/java/com/beyond3/yyGang/cart/Cart.java
@@ -1,0 +1,34 @@
+package com.beyond3.yyGang.cart;
+
+import com.beyond3.yyGang.CartOption;
+import com.beyond3.yyGang.user.Users;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.OneToOne;
+import jakarta.persistence.Table;
+import lombok.Getter;
+
+import java.util.List;
+
+@Entity
+@Getter
+@Table(name = "cart")
+public class Cart {
+
+    @Id
+    @GeneratedValue
+    @Column(name = "cart_id")
+    private Long cartId;
+
+    @OneToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id")
+    private Users user;
+
+    @OneToMany(mappedBy = "cart")
+    private List<CartOption> cartOptions;
+}
