--- yyGang/src/main/java/com/beyond3/yyGang/category/domain/IngredientCategory.java
@@ -1,32 +1,0 @@
-package com.beyond3.yyGang.category.domain;
-
-import com.beyond3.yyGang.store.domain.NSupplements;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import lombok.Getter;
-
-@Entity
-@Getter
-@Table(name = "ingredient_category")
-public class IngredientCategory {
-    @Id
-    @GeneratedValue
-    @Column(name = "i_category_id")
-    private Long iCategoryId;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "products_id")
-    private NSupplements nSupplements;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "ingredient_id")
-    private Ingredient ingredient;
-
-
-}
