+++ yyGang/src/main/java/com/beyond3/yyGang/ingredient/Ingredient.java
@@ -1,0 +1,20 @@
+package com.beyond3.yyGang.ingredient;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.Getter;
+
+@Entity
+@Getter
+@Table(name = "ingredient")
+public class Ingredient {
+    @Id
+    @GeneratedValue
+    @Column(name = "ingredient_id")
+    private Long ingredientID; // 영양 성분 ID
+
+    private String ingredient; // 영양 성분
+}
