+++ yyGang/src/main/java/com/beyond3/yyGang/pay/PayStatus.java
@@ -1,0 +1,5 @@
+package com.beyond3.yyGang.pay;
+
+public enum PayStatus {
+    WAITING, FAIL, SUCCESS
+}
