+++ yyGang/src/main/java/com/beyond3/yyGang/pay/Payment.java
@@ -1,0 +1,46 @@
+package com.beyond3.yyGang.pay;
+
+import com.beyond3.yyGang.order.Order;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.OneToOne;
+import jakarta.persistence.Table;
+import lombok.Getter;
+
+import java.time.LocalDateTime;
+
+@Entity
+@Getter
+@Table(name = "payment")
+public class Payment {
+
+    @Id
+    @GeneratedValue
+    @Column(name = "payment_id")
+    private Long paymentId; // 결제 ID
+
+    private Long totalPrice; // 전체 가격
+
+    private String payMethod; // 결제 방법
+
+    @Enumerated(EnumType.STRING)
+    @Column(name = "pay_status")
+    private PayStatus payStatus;  // 결제 상태 - WAITING, FAIL, SUCCESS
+
+    @Column(columnDefinition = "TIMESTAMP DEFAULT CURRENT_TIMESTAMP")
+    private LocalDateTime date; // 결제 날짜
+
+    @OneToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "order_id")
+    private Order order;  // 주문 아이디 참조
+
+
+
+
+}
