--- yyGang/src/main/java/com/beyond3/yyGang/store/domain/PayStatus.java
@@ -1,5 +1,0 @@
-package com.beyond3.yyGang.store.domain;
-
-public enum PayStatus {
-    WAITING, FAIL, SUCCESS
-}
