+++ yyGang/src/main/java/com/beyond3/yyGang/user/Gender.java
@@ -1,0 +1,5 @@
+package com.beyond3.yyGang.user;
+
+public enum Gender {
+    Male, Female
+}
