+++ yyGang/src/main/java/com/beyond3/yyGang/user/Role.java
@@ -1,0 +1,31 @@
+package com.beyond3.yyGang.user;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.Table;
+import lombok.Getter;
+
+import java.util.List;
+
+@Entity
+@Getter
+@Table(name = "role")
+public class Role {
+
+    @Id
+    @GeneratedValue
+    @Column(name = "role_id")
+    private Long roleId;
+
+    @Enumerated(EnumType.STRING)
+    @Column(name = "role_name")
+    private Role_name roleName;
+
+    @OneToMany(mappedBy = "role")
+    private List<Users> users;
+}
